SQLite INTEGERs can be up to 8-bytes large.

+++ DatabaseLibrary/src/main/java/com/wit/databaselibrary/model/DatabaseObject.java
@@ -1,13 +1,13 @@
 package com.wit.databaselibrary.model;
 
 public class DatabaseObject {
-	private final Integer id;
+	private final Long id;
 
-	public DatabaseObject( final Integer id ) {
+	public DatabaseObject( final Long id ) {
 		this.id = id;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 }
